requirements.txt
psycopg2-binary
pandas
python-decouple
click
tqdm
